--- recruit/src/main/java/net/slipp/franchise/application/service/SampleService.java
@@ -1,4 +1,0 @@
-package net.slipp.franchise.application.service;
-
-public class SampleService {
-}
