--- recruit/src/main/java/net/slipp/franchise/domain/model/recruit/RecruitRepository.java
@@ -1,13 +1,0 @@
-package net.slipp.franchise.domain.model.recruit;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RecruitRepository {
-
-    RecruitId nextId();
-
-    void save(Recruit recruit);
-
-    Recruit findById(RecruitId of);
-}
