--- recruit/src/main/java/net/slipp/franchise/domain/model/recruit/Status.java
@@ -1,11 +1,0 @@
-/**
- * Created by joenggyu0@gmail.com on 4/7/20
- * Github : http://github.com/lenkim
- */
-
-package net.slipp.franchise.domain.model.recruit;
-
-public enum Status {
-
-    BEGIN, START, FINISH
-}
