--- recruit/src/main/java/net/slipp/franchise/presentations/HelloController.java
@@ -1,13 +1,0 @@
-package net.slipp.franchise.presentations;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HelloController {
-
-    @GetMapping("/hello")
-    public String hello() {
-        return "hello world!";
-    }
-}
