+++ recruit/src/main/java/net/slipp/moim/application/service/SampleService.java
@@ -1,0 +1,4 @@
+package net.slipp.moim.application.service;
+
+public class SampleService {
+}
