+++ recruit/src/main/java/net/slipp/moim/domain/model/recruit/Status.java
@@ -1,0 +1,11 @@
+/**
+ * Created by joenggyu0@gmail.com on 4/7/20
+ * Github : http://github.com/lenkim
+ */
+
+package net.slipp.moim.domain.model.recruit;
+
+public enum Status {
+
+    BEGIN, START, FINISH
+}
