+++ recruit/src/main/java/net/slipp/moim/presentations/HelloController.java
@@ -1,0 +1,13 @@
+package net.slipp.moim.presentations;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HelloController {
+
+    @GetMapping("/hello")
+    public String hello() {
+        return "hello world!";
+    }
+}
