--- recruit/src/test/java/net/slipp/franchise/domain/model/DomainTestSupport.java
@@ -1,9 +1,0 @@
-package net.slipp.franchise.domain.model;
-
-import net.slipp.common.domain.model.DomainEventTestSupport;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith(MockitoExtension.class)
-public abstract class DomainTestSupport implements DomainEventTestSupport {
-}
