+++ recruit/src/test/java/net/slipp/moim/domain/model/DomainTestSupport.java
@@ -1,0 +1,9 @@
+package net.slipp.moim.domain.model;
+
+import net.slipp.common.domain.model.DomainEventTestSupport;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+@ExtendWith(MockitoExtension.class)
+public abstract class DomainTestSupport implements DomainEventTestSupport {
+}
